changed resourceCard to be a new component itself
