+++ src/main/app/src/App.js
@@ -3,6 +3,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import Routes from './Routes';
 import Navbar from './constants/Navbar';
 import Line from './constants/Line';
+import ResourceCard from './constants/ResourceCard';
 import info from './constants/keys.js';
 
 /* Component for entire application */
