+++ src/main/app/src/pages/HelpPage.js
@@ -6,19 +6,7 @@ import info from '../constants/keys.js';
 // Import Bootstrap CSS and JS. Only this page uses these. 
 import 'bootstrap/dist/css/bootstrap.css';
 import 'bootstrap/dist/js/bootstrap.js';
-
-// This is a stateless, functional React component used to render each resource in a card format. 
-const ResourceCard = ({ organization, description, url }) => (
-  <div className = "p-3">
-    <div className="card">
-      <div className="card-body">
-        <h5 className="card-title text-info">{organization}</h5>
-        <p className="card-text text-secondary" >{description}</p>
-        <a className="card-text btn btn-light" href={url}>View more</a>
-      </div>
-    </div>
-  </div>
-);
+import ResourceCard from '../constants/ResourceCard';
 
 /* Component for home page */
 const HelpPage = () => {
